Move tests and add animation gating test
